Move config bar from viewer container to pane
