+++ app/src/js/components/label_pane.tsx
@@ -1,13 +1,30 @@
+import { IconButton } from '@material-ui/core'
+import Grid from '@material-ui/core/Grid'
+import MenuItem from '@material-ui/core/MenuItem'
+import Select from '@material-ui/core/Select'
+import CloseIcon from '@material-ui/icons/Close'
+import ViewStreamIcon from '@material-ui/icons/ViewStream'
 import { withStyles } from '@material-ui/styles'
 import * as React from 'react'
 import SplitPane from 'react-split-pane'
-import { updateAll } from '../action/common'
+import { changeViewerConfig, deletePane, splitPane, updateAll } from '../action/common'
 import Session from '../common/session'
-import { resizerStyles } from '../styles/split_pane'
+import * as types from '../common/types'
+import { SplitType } from '../functional/types'
+import { paneBarStyles, resizerStyles } from '../styles/split_pane'
 import { Component } from './component'
 import ViewerContainer, { viewerContainerReactKey } from './viewer_container'
+import { makeDefaultViewerConfig } from '../functional/states'
 
 interface ClassType {
+  /** grid */
+  viewer_container_bar: string
+  /** select */
+  select: string
+  /** icon */
+  icon: string
+  /** icon rotated */
+  icon90: string
   /** class name for resizer */
   resizer: string
 }
@@ -31,11 +48,109 @@ class LabelPane extends Component<Props> {
   public render () {
     const pane = this.state.user.layout.panes[this.props.pane]
     if (pane.viewerId >= 0) {
+      const viewerConfig = this.state.user.viewerConfigs[pane.viewerId]
+      const viewerTypeMenu = (
+        <Select
+          value={viewerConfig.type}
+          onChange={(e) => {
+            const newConfig = makeDefaultViewerConfig(
+              e.target.value as types.ViewerConfigTypeName,
+              viewerConfig.pane,
+              viewerConfig.sensor
+            )
+            if (newConfig) {
+              Session.dispatch(changeViewerConfig(
+                pane.viewerId,
+                newConfig
+              ))
+            }
+          }}
+          classes={{ select: this.props.classes.select }}
+          inputProps={{
+            classes: {
+              icon: this.props.classes.icon
+            }
+          }}
+        >
+          <MenuItem value={types.ViewerConfigTypeName.IMAGE}>Image</MenuItem>
+          <MenuItem value={types.ViewerConfigTypeName.POINT_CLOUD}>
+            Point Cloud
+          </MenuItem>
+          <MenuItem value={types.ViewerConfigTypeName.IMAGE_3D}>
+            Image 3D
+          </MenuItem>
+        </Select>
+      )
+
+      const verticalSplitButton = (
+        <IconButton
+          className={this.props.classes.icon90}
+          onClick={() => {
+            Session.dispatch(splitPane(
+              pane.id,
+              SplitType.VERTICAL,
+              pane.viewerId
+            ))
+          }}
+        >
+          <ViewStreamIcon />
+        </IconButton>
+      )
+
+      const horizontalSplitButton = (
+        <IconButton
+          className={this.props.classes.icon}
+          onClick={() => {
+            Session.dispatch(splitPane(
+              pane.id,
+              SplitType.HORIZONTAL,
+              pane.viewerId
+            ))
+          }}
+        >
+          <ViewStreamIcon />
+        </IconButton>
+      )
+
+      const deleteButton = (
+        <IconButton
+          className={this.props.classes.icon}
+          onClick={() => {
+            Session.dispatch(deletePane(
+              pane.id,
+              pane.viewerId
+            ))
+          }}
+        >
+          <CloseIcon />
+        </IconButton>
+      )
+
+      const viewerContainerBar = (
+          <Grid
+            justify={'flex-end'}
+            container
+            direction='row'
+            classes={{
+              container: this.props.classes.viewer_container_bar
+            }}
+          >
+            {viewerTypeMenu}
+            {verticalSplitButton}
+            {horizontalSplitButton}
+            {deleteButton}
+          </Grid>
+      )
       // Leaf, render viewer container
-      return (<ViewerContainer
-        id={pane.viewerId}
-        key={viewerContainerReactKey(pane.viewerId)}
-      />)
+      return (
+        <div>
+          {viewerContainerBar}
+          <ViewerContainer
+            id={pane.viewerId}
+            key={viewerContainerReactKey(pane.viewerId)}
+          />
+        </div>
+      )
     }
 
     if (
@@ -74,6 +189,9 @@ class LabelPane extends Component<Props> {
   }
 }
 
-const StyledLabelPane = withStyles(resizerStyles)(LabelPane)
+const StyledLabelPane = withStyles((_theme) => ({
+  ...resizerStyles(),
+  ...paneBarStyles()
+}))(LabelPane)
 
 export default StyledLabelPane
