+++ app/src/js/styles/split_pane.tsx
@@ -1,5 +1,30 @@
 import createStyles from '@material-ui/core/styles/createStyles'
 
+export const paneBarStyles = () => createStyles({
+  viewer_container_bar: {
+    position: 'absolute',
+    zIndex: 10
+  },
+  select: {
+    'backgroundColor': 'rgba(34, 34, 34, 1)',
+    'border': '1px solid #ced4da',
+    'color': '#ced4da',
+    'borderRadius': 4,
+    'padding': '10px 26px 10px 12px',
+    '&:focus': {
+      borderRadius: 4
+    },
+    'margin-right': '5px'
+  },
+  icon: {
+    color: '#ced4da'
+  },
+  icon90: {
+    color: '#ced4da',
+    transform: 'rotate(90deg)'
+  }
+})
+
 export const resizerStyles = () => createStyles({
   resizer: {
     'background': 'rgba(150, 150, 150, 1)',
